requirements.txt
# Backend - Django
Django==4.2.30
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
python-dotenv==1.0.0
Pillow==10.1.0

# OCR y procesamiento de imágenes
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
pymupdf==1.23.26

# HTTP requests (Mistral AI client)
requests==2.31.0

# Cache
django-redis==5.4.0

# Tareas asíncronas (Celery)
celery==5.3.6
redis==5.0.1

# Base de datos
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# IA y NLP (para parseo de facturas)
mistralai>=0.1.0

# Utilidades
python-dateutil==2.8.2
tqdm==4.66.1
structlog==23.2.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Seguridad y Similitud de Texto
thefuzz==0.20.0
python-magic==0.4.27

# Error tracking y monitoreo
sentry-sdk==1.40.0

# Rate limiting
django-ratelimit==4.1.0

# Generación de PDF
weasyprint==62.3
pydyf==0.11.0
